requirements.txt
ag-ui-protocol==0.1.15
aiofile==3.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.93.0
anyio==4.13.0
argcomplete==3.6.3
async-timeout==5.0.1
attrs==26.1.0
Authlib==1.6.9
backports.tarfile==1.2.0
beartype==0.22.9
beautifulsoup4==4.14.3
blinker==1.9.0
boto3==1.42.87
botocore==1.42.87
cachetools==7.0.5
caio==0.9.25
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cohere==6.0.0
colorama==0.4.6
cryptography==46.0.7
cyclopts==4.10.2
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
eval_type_backport==0.3.1
exceptiongroup==1.3.1
executing==2.2.1
fastavro==1.12.1
fastmcp==3.2.3
filelock==3.25.2
Flask==3.1.3
frozenlist==1.8.0
fsspec==2026.3.0
genai-prices==0.0.56
google-auth==2.49.2
google-genai==1.72.0
googleapis-common-protos==1.74.0
greenlet==3.4.0
griffelib==2.0.2
groq==1.1.2
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.10.1
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsonpatch==1.33
jsonpath-python==1.1.5
jsonpointer==3.1.1
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
keyring==25.7.0
langchain-core==1.2.28
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.30
logfire==4.31.1
logfire-api==4.31.1
lxml==6.0.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.27.0
mdurl==0.1.2
mistralai==2.3.2
more-itertools==11.0.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.4.2
nexus-rpc==1.4.0
numpy==2.2.6
openai==2.31.0
openapi-pydantic==0.5.1
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-httpx==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
orjson==3.11.8
ormsgpack==1.12.2
packaging==25.0
pathable==0.5.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
py-key-value-aio==0.4.4
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-ai==1.79.0
pydantic-ai-slim==1.79.0
pydantic-evals==1.79.0
pydantic-graph==1.79.0
pydantic-handlebars==0.1.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
PyMuPDF==1.27.2.2
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.25
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.4.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
sse-starlette==3.3.4
starlette==1.0.0
sympy==1.14.0
temporalio==1.25.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.4.1
torch==2.11.0
tqdm==4.67.3
transformers==5.5.3
typer==0.24.1
types-protobuf==6.32.1.20260221
types-requests==2.33.0.20260408
typing-inspection==0.4.2
typing_extensions==4.15.0
uncalled-for==0.3.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
Werkzeug==3.1.8
wrapt==1.17.3
xai-sdk==1.11.0
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
